Feat: add HTTP server with plain text response on port 1245
This commit adds a small http server using the http module in NodeJs.
The server is assigned to the variable 'app' and exported for use in other modules.
The server listens on port 12454 and responds with the message 'Hello Holberton School!' as plain text for any endpoint.The code is al
The code is also documented using JSDoc comments.